+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/Components/ComponentList.swift
@@ -27,29 +27,32 @@ import SwiftUI
 
 struct ComponentsCardsList: View {
     let componentList = [
-        TextButtonComponentModel(name: "Button", image: "Shape_button") {
-            ShapeButtonsList()
-        },
-        TextButtonComponentModel(name: "Text Button", image: "Text_button") {
+        ComponentModel(name: "Text button", image: "Text_button") {
             TextButtonPage()
         },
-        TextButtonComponentModel(name: "Bar", image: "empty") {
-            TabBarPage()
+        ComponentModel(name: "Shape button", image: "Shape_button") {
+            ShapeButtonPage()
         },
-        TextButtonComponentModel(name: "Slider", image: "Controls") {
+        ComponentModel(name: "Slider", image: "Controls") {
             SliderPage()
         },
-        TextButtonComponentModel(name: "Edit Text", image: "empty") {
-            EditTextPage()
+
+        ComponentModel(name: "Card", image: "Card") {
+            CardPage()
         },
-        TextButtonComponentModel(name: "Card", image: "Card") {
-            CardViewDemoSandbox()
-        },
-        TextButtonComponentModel(name: "Progress indicator", image: "empty") {
+        ComponentModel(name: "Progress indicator", image: "Progress_indicator") {
             ProgressIndicatorPage()
         },
+<<<<<<< HEAD
         TextButtonComponentModel(name: "Lists", image: "empty") {
             ListsPage()
+=======
+        ComponentModel(name: "Edit text", image: "empty") {
+            EditTextPage()
+        },
+        ComponentModel(name: "Bar", image: "empty") {
+            TabBarPage()
+>>>>>>> a4e4bb9cdad26bab02471502778770d13c6a2709
         },
     ]
 
@@ -63,7 +66,7 @@ struct ComponentsCardsList: View {
                 Spacer().frame(height: 15)
                 LazyVGrid(columns: columns, spacing: 15) {
                     ForEach(componentList) {
-                        TextButtonComponent(component: $0)
+                        Component(component: $0)
                     }
                 }.padding([.leading, .trailing])
                 Spacer().frame(height: 15)