+++ OrangeDesignSystemDemo/Pages/CardPage.swift
@@ -25,7 +25,7 @@ import Foundation
 import OrangeDesignSystem
 import SwiftUI
 
-struct CardViewDemoSandbox: View {
+struct CardPage: View {
     @State var showImage = false
     @State var showSubtitle = false
     @State var showDescription = false
@@ -50,11 +50,13 @@ struct CardViewDemoSandbox: View {
 
     var body: some View {
         ScrollView {
-            VStack(alignment: .leading, spacing: 15) {
+            VStack(alignment: .leading, spacing: 20) {
 
-                /*** Card View */
+                ComponentDescription(text: "Cards are a contained and independent element that can display content and actions on a single topic.")
+                VariantsTitle()
+
+                // Card demonstrator
                 CardViewCustom(element: example) {
-
                     if showButton {
                         Button {} label: {
                             ODSGenericButtonContent(topText: "Button", textColor: ODSColor.coreBlack.color)
@@ -64,7 +66,7 @@ struct CardViewDemoSandbox: View {
                 }
                 .padding()
 
-                /*** Custom */
+                // Controls
                 VStack {
                     Toggle(isOn: $showImage) {
                         Text("Show image")
@@ -93,13 +95,21 @@ struct CardViewDemoSandbox: View {
                             .frame(maxWidth: .infinity)
                     }
                     .buttonStyle(ODSFilledButtonStyle())
+<<<<<<< HEAD:OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/Components/CardViewDemoSandbox.swift
                 }.padding()
                     .font(ODSFontStyle.bodyRegular.font())
                     .foregroundColor(.primary)
+=======
+                }
+                .padding()
+                .font(ODSFontStyle.bodyRegular.font())
+                .foregroundColor(.primary)
+>>>>>>> a4e4bb9cdad26bab02471502778770d13c6a2709:OrangeDesignSystemDemo/Pages/CardPage.swift
             }
-
-        }.navigationTitle("Card")
-            .background(Color(uiColor: .systemGray5))
+            .padding(EdgeInsets(top: 0, leading: 15, bottom: 5, trailing: 15))
+        }
+        .navigationTitle("Card")
+        .background(Color(uiColor: .systemGray5))
     }
 }
 
@@ -107,12 +117,12 @@ struct CardViewDemoSandbox: View {
 struct CardViewDemoSandBox_Previews: PreviewProvider {
     static var previews: some View {
         NavigationView {
-            CardViewDemoSandbox()
+            CardPage()
                 .previewInterfaceOrientation(.portrait)
         }
 
         NavigationView {
-            CardViewDemoSandbox()
+            CardPage()
                 .previewInterfaceOrientation(.portrait)
                 .environment(\.dynamicTypeSize, .accessibility3)
         }